chore: Organize code
