+++ Sources/Generator/Cached/GMarkCachedManager.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 import UIKit
-import CryptoKit
 
 final class GMarkCachedManager {
     
@@ -56,34 +55,3 @@ final class GMarkCachedManager {
 }
     
 
-// MARK: - Extensions
-
-private extension Data {
-    /// Computes the MD5 hash of the data.
-    ///
-    /// - Returns: A hexadecimal string representation of the MD5 hash.
-    func md5Hash() -> String {
-        let digest = Insecure.MD5.hash(data: self)
-        return digest.map { String(format: "%02hhx", $0) }.joined()
-    }
-}
-
-private extension String {
-    /// Computes the MD5 hash of the string.
-    ///
-    /// - Returns: A hexadecimal string representation of the MD5 hash.
-    func md5() -> String {
-        self.data(using: .utf8)?.md5Hash() ?? ""
-    }
-    
-    func trimmed() -> String {
-        self.trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-    
-    var cacheKey: String {
-        let trimmedKey = self.trimmed()
-        guard !trimmedKey.isEmpty else { return UUID().uuidString }
-        // 使用 MD5 哈希生成唯一的缓存键
-        return trimmedKey.md5()
-    }
-}
